chore: undid changes and created proper test for regexes
This reverts commit 2b10fdc6d6c8db0133d6e43636cc7b7d933f6fe1.

using functional esk test formatting for regex testing

Revert "refactor: moved junit to allprojects"

This reverts commit 91e064dabf5407658e5cf16bb7b1f8fc888e8b14.
